Extract shared Ace editor helpers and document editor reset behavior
